refactor(app): Move services folder to main folder
