+++ src/app/main/services/component-communication-event.service.spec.ts
